Handle `C.TIME_END_OF_SOURCE` buffer timestamps in `CeaDecoder`
The behaviour was changed in 1.4.0 with https://github.com/androidx/media/commit/0f42dd47526c70e38b1048bb18c994c281800485,
so that the buffer timestamp is compared to `outputStartTimeUs` when
deciding whether to discard a "decode only" buffer before decoding
(instead of the deprecated/removed `isDecodeOnly` property). This breaks
when the buffer timestamp is `TIME_END_OF_SOURCE` (which is
`Long.MIN_VALUE`), because `TIME_END_OF_SOURCE < outputStartTimeUs` is
always true, so the end-of-stream buffer is never passed to the decoder
and on to `TextRenderer` where it is used to
[set `inputStreamEnded = true`](https://github.com/androidx/media/blob/40f187e4b47c445af5049a5a49ee4633ce4c1a76/libraries/exoplayer/src/main/java/androidx/media3/exoplayer/text/TextRenderer.java#L434-L436)
and so playback hangs.

#cherrypick

Issue: androidx/media#1863
PiperOrigin-RevId: 695767247