Optionally preserve `DrmSessions` after `MediaSource.releaseSource()`
The `DefaultDrmSessionManager` releases all keep-alive DrmSessions after the
`MediaSource` is released, as such the only way to preserve keep-alive sessions
from one `MediaItem` playback to the next is to use `Playlist`'s.

`Playlist' is not practical for a large live channel set for a couple of
reasons:

1. Keeping the `MediaSource` active consumes considerable memory
2. Polling of the MPD/HLS Playlist continues for all `MediaItems` that have ever played in the playlist.

This change is conceptually quite simple:

1. Preserve the active list of cached `DrmSession` when `DrmSessionManager.release()` called
2. Add a method that can be used on `Activity.onStop()` to release the cached `DrmSession`'s

This new behavior is guarded by a setting that defaults to the old behavior.

This commit also includes a fix to remove `preacquiredSessionReferences` when their associated session is fully
released (to avoid the logged error message for `referenceCount` < 0)