Try input format PCM encoding first in MediaCodecAudioRenderer
...and remove hard requirement for 16-bit PCM support in the audio
sink even if that's not needed to playback.

If the audio file is using 24-bit or 32-bit (int) PCM,
MediaCodecAudioRenderer tries to output float, and if that is not
supported, falls back to 16-bit PCM. This commit changes this
behavior to first trying the media file format, then any close
high-resolution audio formats (such as float) and then fall back
to 16-bit PCM. The behaviour with DefaultAudioSink does not change,
but if an audio sink supports any more optimal formats, they will
now be used with MediaCodecAudioRenderer.