Experiment with adding a third SeekHead
This file was crafted by hand in a hex editor by modifying
`sample.mkv`:
1. Adding two SeekHead elements at the end, each with a single Seek
   entry, with the first pointing at the Cues element and the second
   pointing at the Tags element.
2. Update the original SeekHead element so the Seek entries that
   previously pointed to the Cues and Tags element now point out the two
   new SeekHead elements we added.

The resulting file is successfully parsed by `mkvinfo` and playable with
VLC.

The dump files were generated with all changes in MatroskaExtractor
reverted (which is why they're unseekable).

Bringing the MatroskaExtractor changes back and running the test results
in a larger than expected diff in the dump file. Specifically all the
samples are dropped.